REFACTOR test classes into `Feature` & `Unit` namespaces
